+++ CHANGELOG.md
@@ -17,7 +17,9 @@ Security in case of vulnerabilities.
 ## Unreleased
 
 ### Added
--Land registry title plans` task for converson project.
+- Land registry title plans` task for conversion project.
+- Added `Master Funding Agreement` task for both conversion and transfer projects.
+- Added `Incoming Trust CEO contact` task page
 
 ### Changed
 
@@ -70,6 +72,10 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - Fixed broken task note urls.
 - Fixed all project by month query for conversion projects if all conditions met is null
 
+### Security
+
+- Fixed reflected Cross-Site Scripting (XSS) vulnerability on cookies page (150001) - added server-side URL validation to prevent malicious script injection via returnUrl parameter
+
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-25.771...production-2025-10-03.811) for everything in the release
 
 ---
@@ -90,6 +96,8 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 
 ### Changed
 - Group reference number links to the group on "About the project"
+- Allow users to modify `declaration of expenditure certificate date` on the `Receive declaration of expenditure certificate` task for both conversion and transfer project.
+
 
 ### Fixed
 - GroupReferenceNumberAttribute failed when there was no existing group